Change dtype of arrays for temporaries array

+++ psydac/feec/pushforward.py
@@ -285,7 +285,7 @@ class Pushforward:
         if isinstance(space, (VectorFemSpace, ProductFemSpace)):
 
             fields_to_push = [np.ascontiguousarray(fields_eval[i][index_trim[i]]) for i in range(self.ldim)]
-            pushed_fields_list = [np.zeros_like(fields_to_push[i]) for i in range(self.ldim)]
+            pushed_fields_list = [np.zeros_like(fields_to_push[i], dtype=fields_eval.dtype) for i in range(self.ldim)]
             if self.ldim == 2:
                 for i in range(2):
                     pushforward_2d_l2(fields_to_push[i], self.sqrt_metric_det_temp, pushed_fields_list[i])
@@ -327,7 +327,7 @@ class Pushforward:
 
         fields_eval = np.ascontiguousarray(np.stack([fields_eval[i][(*index_trim[i], Ellipsis)] for i in range(self.ldim)], axis=0))
 
-        pushed_fields = np.zeros((fields_eval.shape[-1], *fields_eval.shape[:-1]))
+        pushed_fields = np.zeros((fields_eval.shape[-1], *fields_eval.shape[:-1]), dtype=fields_eval.dtype)
         if self.ldim == 2:
             pushforward_2d_hdiv(fields_eval, self.jac_temp, self.sqrt_metric_det_temp, pushed_fields)
         if self.ldim == 3:
@@ -350,7 +350,7 @@ class Pushforward:
         if self.inv_jac_temp is None:
             self.inv_jac_temp = self.jacobian_inv()
         fields_eval = np.ascontiguousarray(np.stack([fields_eval[i][(*index_trim[i], Ellipsis)] for i in range(self.ldim)], axis=0))
-        pushed_fields = np.zeros((fields_eval.shape[-1], *fields_eval.shape[:-1]))
+        pushed_fields = np.zeros((fields_eval.shape[-1], *fields_eval.shape[:-1]), dtype=fields_eval.dtype)
 
         if self.ldim == 2:
             pushforward_2d_hcurl(fields_eval, self.inv_jac_temp, pushed_fields)
